requirements.txt
streamlit
groq
google-genai
streamlit-mic-recorder
gTTS

tomli
